requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==5.4.1
pytesseract==0.2.6
Werkzeug==0.14.1
python-slugify==2.0.1
gunicorn==19.9.0